% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_treated}
\alias{plot_treated}
\title{Plot treatment effects and uncertainty intervals for all backends}
\usage{
plot_treated(fits, level = 0.95, start = 1, compare = NULL, ylim = NULL)
}
\arguments{
\item{fits}{List of fits, e.g. from calling \code{\link[=run_sc]{run_sc()}}.}

\item{level}{Level for plotted confidence / credible intervals.}

\item{start}{Index of first time to plot. Times shown on the x axis will range
from time \code{start} to the last time.}

\item{compare}{If equal to one of "gsynth", "bpcausal", or "stan", the plot will
emphasize the inferred effects from backend \code{compare}, allowing for easier
comparison between method \code{compare} and others.}

\item{ylim}{If equal to a vector of two numbers, defines the lower and upper y axis
limits.}
}
\description{
\code{plot_treated()} plots the estimated treatment effects against the time variable
for all treated unit and all fitted backends. Point estimates are plotted along with
uncertainty intervals. For gsynth, uncertainty intervals are bootstrap confidence
intervals. For bpCausal and stan, the point estimates are posterior means, and the
uncertainty intervals are credible intervals.
}
