% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sc.R
\name{run_sc}
\alias{run_sc}
\title{Generate synthetic control estimates from specified backends}
\usage{
run_sc(
  data,
  response,
  panel_index,
  covars,
  num_latent,
  treated_index,
  bpc_warmup = 5000,
  bpc_iter = 15000,
  stan_warmup = 1000,
  stan_iter = 2000,
  stan_ad = 0.8,
  stan_mt = 11,
  include_covars = c("gs", "bp", "st"),
  backends = c("gs", "bp", "st")
)
}
\arguments{
\item{data}{The input panel data in long format, which
must include, at minimum, columns for the response variable,
unit variable, time variable, and treatement indicator variable.}

\item{response}{Name of column containing the response variable.}

\item{panel_index}{Vector of two strings containing (i) the name
of the column containing the unit variable, and (ii) the name
of the column containing the time variable.}

\item{covars}{Vector of names of columns corresponding to regression
covariate variables which are adjusted for.}

\item{num_latent}{The number of latent factors to fit (used for both
bpCausal and Stan models). For bpCausal, some number of factors less
than num_latent may be selected by the model's specification search.}

\item{bpc_warmup}{Number of warmup / burn-in iterations to run for
bpCausal's sampler.}

\item{bpc_iter}{Number of post-warmup iterations to run for bpCausal's
sampler.}

\item{stan_warmup}{Number of warmup / burn-in iterations to run for
Stan's sampler.}

\item{stan_iter}{Number of post-warmup iterations to run for Stan's
sampler.}

\item{stan_ad}{Adapt-delta hyperparameter for Stan's dHMC sampler.}

\item{stan_mt}{Max-treedepth hyperparameter for Stan's dHMC sampler.}

\item{include_covars}{Vector of backends for which covariate variables
should be adjusted for.}

\item{backends}{Vector of backends for which to perform inference.}
}
\value{
List with entries \code{gs_fit}, \code{bp_fit}, and \code{st_fit} containing
inference output from gsynth, bpCausal, and Stan methods respectively,
or \code{NULL} if corresponding method not included in \code{backends}.
}
\description{
\code{run_sc()} estimates causal effects for treated units in
panel data using three variants of the synthetic control
idea: (i) the generalized synthetic control estimator
(implemented by the gsynth package), (ii) the DF-LFM
Bayesian latent factor model, and (iii) a custom variant of
the DF-LFM model implemented in Stan.
}
