% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wg_example.R
\name{run_wg_example}
\alias{run_wg_example}
\title{Run backends on West German reunification data}
\usage{
run_wg_example(refit = FALSE)
}
\arguments{
\item{refit}{Should the synthetic control estimates be recomputed from each
available backend.}
}
\value{
List of inference outputs from each of the three available backend
methods.
}
\description{
Runs all three backends on the classic West German reunification example.
Synthetic control estimates are constructed for the effect of reunification
on the West German GDP. By default, estimates are loaded in from pre-computed
data, but analysis can be optionally rerun.
}
