% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sc.R
\name{summarize}
\alias{summarize}
\title{Generate inferential summaries from synthetic control fits.}
\usage{
summarize(fits, stats = c("y.ct", "eff", "err"), format = "long", level = NULL)
}
\arguments{
\item{fits}{List of outputs from synthetic control backends, e.g. as output by \code{\link[=run_sc]{run_sc()}}.}

\item{stats}{Vector of inferential summaries to generate. Possible quantites are \code{y.ct} for
untreated potential outcomes of treated units, \code{eff} for causal effects, \code{err} for measures
of error (standard errors for gsynth and posterior SDs for bpCausal and Stan).}

\item{format}{Either "long" to return summaries in long (i.e. data frame) format, or
"wide" to return summaries as a list of arrays.}

\item{level}{Level to use for uncertainty intervals (confidence intervals for gsynth and
credible intervals for bpCausal and Stan). If \code{NULL}, intervals are not generated.}
}
\value{
If \code{format = "long"}, returns data frame with columns for the treated unit,
time index, backend method, and each computed summary quantity. If \code{format = "wide"}, returns
list of arrays for each backend method with extents (number of times) x
(number of treated units) x (number of summary quantities). Note: uncertainty intervals
are only available in long format.
}
\description{
Given a list of synthetic control method fits, \code{summarize()} generates point estimates
for treatment effects and untreated potential outcomes, along with measures
of uncertainty and intervals.
}
